% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron.R
\name{air_triangle}
\alias{air_triangle}
\title{Air d'un triangle( formule heron)}
\usage{
air_triangle(a, b, c)
}
\arguments{
\item{a}{la longueur du premier côte du traingle}

\item{b}{la longueur du deuxième côte du traingle}

\item{c}{la longueur du troisième côte du traingle}
}
\value{
renvoie l'air d'un triangle en utilisant la formule d'héron si les paramètre sont numeric
et supérieur à zéro sinon renvoie un message d'erreur
}
\description{
Cette fonction permet de retouner l'air d'un traingle en utilisant la formule d'héron
}
\examples{
a=3;b=6;c=7
air_triangle(a,b,c)

a="d"; b=1;C=1
air_triangle(a,b,c)

}
