% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcEpitopePValues}
\alias{calcEpitopePValues}
\title{Calculate Epitope-level p-values}
\usage{
calcEpitopePValues(
  probe_pvalues,
  epitope_ids,
  method = "min",
  do.sort = FALSE,
  data_matrix = NULL,
  probes = rownames(probe_pvalues)
)
}
\arguments{
\item{probe_pvalues}{matrix of probe-level p-values}

\item{epitope_ids}{vector of epitope ids to calculate p-values from
(needs to be a unique set)}

\item{method}{meta p-value method to use (see calcProteinValues)}

\item{do.sort}{sort the p-value after calculation?}

\item{data_matrix}{matrix of values to calculate the co-variances for some
meta p-value methods (ebrown and kosts)}

\item{probes}{speed up parameter}
}
\value{
matrix of epitope p-values where each row is an epitope id
}
\description{
Calculate Epitope-level p-values
}
