% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcEpitopePValuesMat}
\alias{calcEpitopePValuesMat}
\title{Calculate epitope-level p-values}
\usage{
calcEpitopePValuesMat(probe_pvalues, epitope_ids, method = "maxFDR")
}
\arguments{
\item{probe_pvalues}{matrix of probe p-values}

\item{epitope_ids}{vector of epitope ids}

\item{method}{meta p-value method to use}
}
\value{
matrix of epitope-level p-values
}
\description{
Calculate epitope-level p-values
}
