% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcEpitopePValuesPDS}
\alias{calcEpitopePValuesPDS}
\title{Title}
\usage{
calcEpitopePValuesPDS(
  probe_pds,
  epitope_ids,
  metap_method = "wmax1",
  p_adjust_method = "BH"
)
}
\arguments{
\item{p_adjust_method}{}
}
\description{
Title
}
