% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcEpitopePValuesProbeDS}
\alias{calcEpitopePValuesProbeDS}
\title{Calculate epitope-level p-values}
\usage{
calcEpitopePValuesProbeDS(
  probe_pds,
  epitope_ids,
  metap_method = "wmax1",
  p_adjust_method = "BH"
)
}
\arguments{
\item{probe_pds}{HERONProbeDataSet with the "pvalue" assay}

\item{epitope_ids}{vector of epitope ids}

\item{metap_method}{meta p-value method to use}

\item{p_adjust_method}{what p.adjust method to use.}
}
\value{
HERONEpitopeDataSet
}
\description{
Calculate epitope-level p-values
}
\examples{
data(heffron2021_wuhan)
pval_seq_res <- calcCombPValues(heffron2021_wuhan)
pval_pr_res <- convertSequenceDSToProbeDS(pval_seq_res)
calls_res <- makeProbeCallsPDS(pval_pr_res)
segments_res <- findEpitopeSegmentsPDS(calls_res, "unique")
epval_res <- calcEpitopePValuesProbeDS(calls_res, segments_res)
}
