% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{calcMetaPValuesMat}
\alias{calcMetaPValuesMat}
\title{Calculate meta p-values on a matrix (column-by-column)}
\usage{
calcMetaPValuesMat(pvalues_mat, by_list, method = "min")
}
\arguments{
\item{pvalues_mat}{matrix of p-values where each row is a feature and
each column is a sample}

\item{by_list}{list of grouping elements (see aggregate)}

\item{method}{what meta p-value method to use.}
}
\value{
matrix of meta p-values
}
\description{
Calculate meta p-values on a matrix (column-by-column)
}
