% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcMetaPValuesVec}
\alias{calcMetaPValuesVec}
\title{Calculate meta p-values given a vector of lower-level p-values}
\usage{
calcMetaPValuesVec(
  pvalues,
  by_list,
  method = "min_bonf",
  do.sort = FALSE,
  data_matrix = NULL
)
}
\arguments{
\item{pvalues}{named vector of probes}

\item{by_list}{list of groupings for meta-pvalue calculation}

\item{method}{meta p-value method to use}

\item{do.sort}{sort in increasing order?}

\item{data_matrix}{matrix of values for Kost's and EBROWN's method}
}
\value{
vector of meta p-values
}
\description{
Calculate meta p-values given a vector of lower-level p-values
}
