% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcMinFDR}
\alias{calcMinFDR}
\title{Calculate minimum FDRs across samples}
\usage{
calcMinFDR(fdrs, additional_stats = TRUE, sort = TRUE)
}
\arguments{
\item{sort}{}
}
\value{

}
\description{
This code calculates the minimum FDR threshold needed to be a hit in X out of N samples
Where samples are in the columns
When additional stats is selected, also calculates the mean and median
}
