% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProbePValuesProbeMat}
\alias{calcProbePValuesProbeMat}
\title{Calculate Probe-level p-values}
\usage{
calcProbePValuesProbeMat(
  probe_mat,
  colData,
  t_sd_shift = NA,
  t_abs_shift = NA,
  t_paired = FALSE,
  z_sd_shift = 0,
  use = "both",
  p_adjust_method = "BH"
)
}
\arguments{
\item{probe_mat}{numeric matrix, rows as seqs and columns as samples}

\item{colData}{experiment design data.frame}

\item{t_sd_shift}{sd shift multiplier for diff t-test}

\item{t_abs_shift}{abs shift to use for the diff t-test}

\item{t_paired}{do paired t-test}

\item{z_sd_shift}{sd shift multiplier for global z-test}

\item{use}{which p-value method(s) to use}

\item{p_adjust_method}{method for adjusting p-values}
}
\value{
matrix of calculated p-values
}
\description{
calculates p-values on the matrix (can be sequence as well), using either a
z-score global, t-stat differential, or a combination of both.
input a matrix that has samples on the columns and sequence probes on the
rows.
}
