% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProbePValuesSeqMat}
\alias{calcProbePValuesSeqMat}
\title{Calculate probe p-values using a sequence matrix.}
\usage{
calcProbePValuesSeqMat(
  seq_mat,
  probe_meta,
  colData,
  t_sd_shift = NA,
  t_abs_shift = NA,
  t_paired = FALSE,
  z_sd_shift = 0,
  use = "both",
  p_adjust_method = "BH"
)
}
\arguments{
\item{seq_mat}{matrix of values where the rows are sequence identifiers
and the columns are samples}

\item{probe_meta}{data.frame with the columns PROBE_ID and PROBE_SEQUENCE}

\item{colData}{design matrix}

\item{t_sd_shift}{standard deviation shift for differential test}

\item{t_abs_shift}{absolute shift for differential test}

\item{t_paired}{run paired analysis}

\item{z_sd_shift}{standard deviation shift for global test}

\item{use}{use global-test ("z"), differential-test ("t"), or both ("both")}

\item{p_adjust_method}{p-value adjustment method to use (default BH)}
}
\value{
matrix of adjusted p-values with additional attributes.
}
\description{
Calculate probe p-values using a sequence matrix.
}
