% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProbePValuesZ}
\alias{calcProbePValuesZ}
\title{Calculate Global p-values Using Normal (z) Distribution}
\usage{
calcProbePValuesZ(probe_mat, pData, sd_shift = 0, all = FALSE)
}
\arguments{
\item{probe_mat}{numeric matrix or data.frame of values}

\item{pData}{design data.frame}

\item{sd_shift}{standard deviation shift to use when calculating p-values
(default 0)}

\item{all}{calculate p-values for all samples, otherwise report the p-values
for just the post samples}
}
\value{
matrix of "p-values"
}
\description{
Calculate Global p-values Using Normal (z) Distribution
}
\examples{
data(heffron2021_wuhan)
probe_meta <- attr(heffron2021_wuhan, "probe_meta")
pData <- attr(heffron2021_wuhan, "pData")
pval_res <- calcProbePValuesZ(heffron2021_wuhan, pData)
}
