% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProteinPValuesE}
\alias{calcProteinPValuesE}
\title{Title}
\usage{
calcProteinPValuesE(
  pvalues_mat,
  method = "min",
  data_matrix = NULL,
  epitopes = rownames(pvalues_mat),
  proteins = getEpitopeProtein(epitopes)
)
}
\arguments{
\item{proteins}{}
}
\value{

}
\description{
Title
}
\examples{
Gets p-values for epitopes, using the protein p-value methods.
Assumes that every row in the matrix is an epitope of the form
protein_start_stop, where start is the first probe, and stop is the last probe in the epitope within the protein.
}
