% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProteinPValuesEDS}
\alias{calcProteinPValuesEDS}
\title{Title}
\usage{
calcProteinPValuesEDS(
  epitope_ds,
  metap_method = "wmin1",
  p_adjust_method = "BH"
)
}
\arguments{
\item{p_adjust_method}{}
}
\description{
Title
}
