% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProteinPValuesMat}
\alias{calcProteinPValuesMat}
\title{Calculate protein-level p-values}
\usage{
calcProteinPValuesMat(epitope_pvalues_mat, method = "maxFDR")
}
\arguments{
\item{epitope_pvalues_mat}{matrix of epitope-level p-values}

\item{method}{meta p-value method to use}
}
\value{
matrix of protein-level p-values
}
\description{
Calculate protein-level p-values
}
