% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProteinPValuesVec}
\alias{calcProteinPValuesVec}
\title{Calculate protein p-values given a vector of probe-level p-values}
\usage{
calcProteinPValuesVec(
  pvalues,
  method = "min_bonf",
  do.sort = FALSE,
  data_matrix = NULL,
  probes = names(pvalues),
  proteins = getProteinLabel(probes)
)
}
\arguments{
\item{pvalues}{named vector of probes/eptiopes}

\item{method}{meta p-value method to use}

\item{do.sort}{sort in increasing order?}

\item{data_matrix}{matrix of values for Kost's and EBROWN's method}

\item{probes}{probe ids on vector}

\item{proteins}{protein labels on vector}
}
\value{
vector of protein-level p-values
}
\description{
Calculate protein p-values given a vector of probe-level p-values
}
