% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosts_fast.R
\name{calculateKostCovarianceFast}
\alias{calculateKostCovarianceFast}
\title{Calculate Kost Covariance}
\usage{
calculateKostCovarianceFast(data_matrix)
}
\arguments{
\item{data_matrix}{An m x n data matrix with each of m rows representing a variable and each of n columns representing a sample. Should be of type numeric matrix.
Note: Method does not deal with missing values.}
}
\value{
An m x m matrix of pairwise covariances between the data vectors calculated using Kost's polynomial fit and the pearson correlation function.
}
\description{
Calculate Kost Covariance
}
