% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catalog.R
\name{Catalog}
\alias{Catalog}
\title{A function for doing simple network analysis on sets of homologs.}
\usage{
Catalog(OrthologyObject, Verbose = FALSE)
}
\arguments{
\item{OrthologyObject}{a matrix of lists, the upper triange of which is filled with matrices of putative homologs}
}
\description{
Catalog takes in an object of type `Orthology` and builds a list of all complete linkage clusters. This list is in the form of matrices that denote how transitive the predicted clusters are.
}
\examples{
Catalog()
}
\keyword{Homology}
