% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{catSequences}
\alias{catSequences}
\title{Concatenate sequences together based upon their start positions.}
\usage{
catSequences(positions, sequences, debug = FALSE)
}
\arguments{
\item{debug}{}
}
\value{
concatenated sequence (character)
}
\description{
Concatenate sequences together based upon their start positions.
}
