% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convertSequenceMatToProbeMat}
\alias{convertSequenceMatToProbeMat}
\title{Convert a sequence matrix to a probe matrix}
\usage{
convertSequenceMatToProbeMat(seq_mat, probe_meta)
}
\arguments{
\item{seq_mat}{matrix with rows as sequences and the columns as samples}

\item{probe_meta}{data.frame with the PROBE_SEQUENCE, PROBE_ID columns}
}
\value{
matrix where the rows are probe ids and the columns are samples
}
\description{
Convert a sequence matrix to a probe matrix
}
\examples{
data(heffron2020_wuhan)
probe_meta <- attr(heffron2020_wuhan, "probe_meta")
probe_mat = convertSequenceMatToProbeMat(heffron2020_wuhan, probe_meta)
}
