% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreAligner.R
\name{CoreAligner}
\alias{CoreAligner}
\title{A function for aligning core genomes.}
\usage{
CoreAligner(ListOfSets, PATH, GeneCalls, IDs, Verbose = FALSE)
}
\arguments{
\item{ListOfSets}{A list of matrices}

\item{PATH}{A character string identifying a sqlite database}

\item{GeneCalls}{A list of dataframes, one for each genome, with named columns "Start", "Stop", and "Strand"}

\item{IDs}{A vector of characters indicating the identifiers for which genomes to select}
}
\description{
Core Aligner is a wrapper for `DECIPHER`'s `AlignTranslation` function. Given a subset of the output of `Catalog` or `SubCatalog`, it will align genes from the complete linkage clusters and then concatenate those alignments.
}
\examples{
CoreAligner()
}
\keyword{Core}
\keyword{Genome}
