% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demi_parametre.R
\name{demi_perimetre}
\alias{demi_perimetre}
\title{Calcule le demi-périmètre d'un triangle}
\usage{
demi_perimetre(a, b, c)
}
\arguments{
\item{a}{Le premier côté du triangle}

\item{b}{Le deuxième côté du triangle}

\item{c}{Le troisième côté du triangle}
}
\value{
Le demi-périmètre du triangle
}
\description{
Calcule le demi-périmètre d'un triangle
}
\examples{
demi_perimetre(3,4,5)
}
