% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egret_report.R
\name{egret_report}
\alias{egret_report}
\title{Generate Rough Report of `EGRET` Outputs}
\usage{
egret_report(
  eList_estim = NULL,
  eList_series = NULL,
  out_path = file.path(getwd(), "EGRET_report.pdf")
)
}
\arguments{
\item{eList_estim}{(list) list returned by `EGRET::modelEstimation`}

\item{eList_series}{(list) list returned by `EGRET::runSeries`}

\item{out_path}{(character) name to save report as (must be a PDF) and (optionally) the path to export (defaults to `getwd()`)}
}
\description{
Accepts the lists returned by `EGRET::modelEstimation` and `EGRET::runSeries` and generates a PDF report containing several coarse plots generated from `EGRET` functions for dealing with these lists. Useful for quick assessment of broad patterns in response but more finely-constructed plots should be built elsewhere
}
