% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egret_trends.R
\name{egret_trends}
\alias{egret_trends}
\title{Describe Flow-Normalized Trends in Concentration and Flux}
\usage{
egret_trends(eList_series = NULL, flux_unit = 8)
}
\arguments{
\item{eList_series}{(list) list returned by `EGRET::runSeries`. Must include a dataframe called "Daily" with a column called "Date"}

\item{flux_unit}{(numeric) object of fluxUnit class. See `EGRET::printFluxUnitCheatSheet()` for allowed integer codes}
}
\value{
(dataframe) object containing flow normalized trends for both concentration and flux (see "Metric" column to differentiate)
}
\description{
Accepts the list returned by `EGRET::runSeries` and runs `EGRET::tableChangeSingle` for concentration and flux. Fixes column names to avoid characters that will be lost on export to CSV and binds rows to export as a single object. See the "Metric" column in the returned object to identify which rows are concentration and which are flux
}
