% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{findBlocksT}
\alias{findBlocksT}
\title{Find consecutive probes}
\usage{
findBlocksT(protein.df, protein_tiling)
}
\arguments{
\item{protein.df}{data.frame with the Protein and Starting position of the
probe}

\item{protein_tiling}{tiling for information for each protein}
}
\value{
data.frame with the Protein, Start, Stop, and Number.Of.Probes
columns
}
\description{
Find consecutive probes
}
