% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{findEpitopeSegments}
\alias{findEpitopeSegments}
\title{Find Epitopes from probe stats and calls.}
\usage{
findEpitopeSegments(
    probe_pvalues_res,
    probe_calls,
    segment.method = "unique",
    segment.score.type = "binary",
    segment.dist.method = "hamming",
    segment.cutoff = "silhouette"
)
}
\arguments{
\item{probe_pvalues_res}{results from calcProbePValuesProbeMat}

\item{probe_calls}{results from makeProbeCalls}

\item{segment.method}{which epitope finding method to use}

\item{segment.score.type}{which type of scoring to use for probes
(binary or zscore, applies for hclust or skater)}

\item{segment.dist.method}{what kind of distance score method to use
(hamming, euclidean, ..., applies for hclust or skater)}

\item{segment.cutoff}{for clustering methods, what cutoff to use
(either numeric value or 'silhouette')}
}
\value{
a vector of epitope identifiers or segments found.
}
\description{
Find Epitopes from probe stats and calls.
}
\examples{
data(heffron2020_wuhan)
probe_meta <- attr(heffron2020_wuhan, "probe_meta")
pData <- attr(heffron2020_wuhan, "pData")
pval_res <- calcProbePValuesSeqMat(heffron2020_wuhan, probe_meta, pData)
calls_res <- makeProbeCalls(pval_res)
segments_res = findEpitopeSegments(probe_calls = calls_res);
}
