% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{findEpitopeSegmentsPDS}
\alias{findEpitopeSegmentsPDS}
\title{Title}
\usage{
findEpitopeSegmentsPDS(
  PDS_obj,
  segment_method,
  segment_score_type,
  segment_dist_method,
  segment_cutoff
)
}
\arguments{
\item{PDS_obj}{HERONProbeDataSet with pvalues and calls in the assay}

\item{segment_cutoff}{}
}
\value{
a vector of epitope identifiers or segments found
}
\description{
Title
}
