% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{findEpitopeSegmentsUnique}
\alias{findEpitopeSegmentsUnique}
\title{Find unique set of epitope regions across all sample calls.}
\usage{
findEpitopeSegmentsUnique(probe_sample_calls)
}
\arguments{
\item{probe_sample_calls}{probe hit matrix}
}
\value{
vector of epitope seqments
}
\description{
Find unique set of epitope regions across all sample calls.
}
