% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope.R
\name{getBlockMetricStats}
\alias{getBlockMetricStats}
\title{Extract aggregate metric statistics from blocks (epitopes)}
\usage{
getBlockMetricStats(
  blocks,
  probes,
  metric,
  label,
  proteins = getProteinLabel(probes),
  pos = getProteinStart(probes)
)
}
\arguments{
\item{blocks}{data.frame of with Protein, Start, and Stop columns}

\item{probes}{list of all probes in an experiment}

\item{metric}{numeric.vector of values to collect stats on}

\item{label}{metric label to use in returned values}

\item{proteins}{getProteinLabel(probes) - cached results}

\item{pos}{getProteinStart(probes) - cached results}
}
\value{
data.frame with the min, max, and mean value for each block
extracted from the metric
}
\description{
For each block a
}
