% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{getClusterSegments}
\alias{getClusterSegments}
\title{Get the Cluster Segments from a starting list of epitopes}
\usage{
getClusterSegments(
  all_epitopes,
  sample_probes,
  do.plot = FALSE,
  method = "hclust",
  dist.method = "orig",
  cutoff = "silhouette",
  overlap_cluster_df
)
}
\arguments{
\item{all_epitopes}{list of epitopes to further segment}

\item{sample_probes}{from makeProbeCalls(...), logical matrix of probe calls}

\item{do.plot}{make supplementary plots?}

\item{method}{clustering method to use (hclust, skater, all)}

\item{dist.method}{for clustering methods, what distance metric to use
("hamming", "euclidean")}

\item{cutoff}{what clustering cutoff to use (silhouette or numeric value)}

\item{overlap_cluster_df}{result from getOverlapClusters}
}
\value{
list of results
overlap_clusters - overlap_cluster_df parameter
segments - list of epitope identifiers of the segments
}
\description{
Get the Cluster Segments from a starting list of epitopes
}
