% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{getClusterSegmentsAll}
\alias{getClusterSegmentsAll}
\title{Get all possible cluster segments
determines all possible segmentation regions from a cluster/epitopeid
Uses sample_probes_sub to find the appropriate tiling for the cluster to use.}
\usage{
getClusterSegmentsAll(sample_probes_sub, cluster_id)
}
\arguments{
\item{sample_probes_sub}{TODO}

\item{cluster_id}{TODO}
}
\value{
vector of epitope identifiers
}
\description{
Get all possible cluster segments
determines all possible segmentation regions from a cluster/epitopeid
Uses sample_probes_sub to find the appropriate tiling for the cluster to use.
}
