% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{getClusterSegmentsSkater}
\alias{getClusterSegmentsSkater}
\title{Title}
\usage{
getClusterSegmentsSkater(
  sample_probes_sub,
  cluster_id,
  dist_method = "euclidean",
  cutoff = "silhouette",
  debug = FALSE
)
}
\arguments{
\item{sample_probes_sub}{TODO}

\item{cluster_id}{TODO}

\item{dist_method}{TODO}

\item{cutoff}{TODO}

\item{debug}{TODO}
}
\value{
vector of segments
}
\description{
Acceptable dist.methods are "euclidean", "hamming"
   "maximum", "manhattan", "canberra", "binary", "minkowski"
 Could also pass in the function, but I haven't tested that yet.
Also "mahalanobis" requires a covariance matrix,
 which I have implemented/tried yet...
Minkowski is hardcoded to have a power (p) of 0.5,
 maybe we should give the user
an option to modify that?
}
