% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{getEpitopeCallsUnique}
\alias{getEpitopeCallsUnique}
\title{Find epitopes using the unique method}
\usage{
getEpitopeCallsUnique(
  probe_sample_padj,
  probe_cutoff = 0.05,
  one_hit_filter = TRUE,
  probes = rownames(probe_sample_padj),
  proteins = getProteinLabel(probes),
  pos = getProteinStart(probes),
  protein_tiling = getProteinTiling(probes)
)
}
\arguments{
\item{probe_sample_padj}{adjusted p-value matrix on the probe-level}

\item{probe_cutoff}{cutoff to use when calling probes for epitope finding}

\item{one_hit_filter}{use one hit filter to filter out epitopes}

\item{probes}{rownames of the probe_sample_padj}

\item{proteins}{proteins of the probes}

\item{pos}{starting position of the probe in the associated protein}

\item{protein_tiling}{tiling of the probe}
}
\value{
list of results
   sample_epitopes - epitopes found per sample (column) as list
   sample_epitopes_fdr - matrix of estimated FDRs per sample for
   sample_epitopes_minfdr = minFDRs;
   all_epitopes - unique vector of epitope ids
   k_of_n_epitopes
}
\description{
This function makes epitope level calls using an FDR matrix.
For each epitope, we calculate the maximum FDR of the probes within that
epitope and use that as the "FDR" for the epitope.  Just means, what
would the probe FDR cutoff need to be in
order to include this epitope for this sample.
}
