% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope.R
\name{getEpitopeID}
\alias{getEpitopeID}
\title{Create EpitopeID from protein, first and last probes}
\usage{
getEpitopeID(protein, start, stop)
}
\arguments{
\item{protein}{vector of proteins}

\item{start}{vector of first probe protein start positions}

\item{stop}{vector of last probe protein start positions}
}
\value{
vector of epitope ids
}
\description{
Create EpitopeID from protein, first and last probes
}
\examples{
getEpitopeID("A", 1, 2)
}
