% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope.R
\name{getEpitopeIDsToProbeIDs}
\alias{getEpitopeIDsToProbeIDs}
\title{Title}
\usage{
getEpitopeIDsToProbeIDs(epitope_ids, tiling = 1)
}
\arguments{
\item{epitope_ids}{vector of epitope identifiers}

\item{tiling}{tling of probes across proteins}
}
\value{
data.frame of epitope_to_probe mappings
}
\description{
Title
}
