% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope.R
\name{getEpitopeProbeIDs}
\alias{getEpitopeProbeIDs}
\title{Get the vector of probes from an epitope id}
\usage{
getEpitopeProbeIDs(epitope_id, tiling = 1)
}
\arguments{
\item{epitope_id}{EpitopeID to obtain probes from}

\item{tiling}{Tiling of the probes across the protein (default 1)}
}
\value{
vector of probe_ids that are contained within the epitope
}
\description{
Get the vector of probes from an epitope id
}
