% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope.R
\name{getEpitopeStart}
\alias{getEpitopeStart}
\title{Obtain first probe's protein start position from Epitope ID}
\usage{
getEpitopeStart(epitope_ids)
}
\arguments{
\item{epitope_ids}{vector of epitope ids}
}
\value{
vector of integers indicating first probe start positions in the
epitope(s)
}
\description{
Obtain first probe's protein start position from Epitope ID
}
\examples{
getEpitopeStart("Prot1_1_5")
}
