% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope.R
\name{getEpitopeStop}
\alias{getEpitopeStop}
\title{Obtain last probe's protein start position from EpitopeID}
\usage{
getEpitopeStop(epitope_ids)
}
\arguments{
\item{epitope_ids}{vector of epitope ids}
}
\value{
vector of integers indicating the last probe protein start position
}
\description{
Obtain last probe's protein start position from EpitopeID
}
\examples{
getEpitopeStop("Prot1_1_5")
}
