% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{getHClustDistMat}
\alias{getHClustDistMat}
\title{Calculate a distance matrix for use with hclust}
\usage{
getHClustDistMat(sample_probes_sub, dist.method)
}
\arguments{
\item{sample_probes_sub}{matrix of calls}

\item{dist.method}{distance method to use (see dist)}
}
\value{
matrix that can be converted to a distance for use with complete
hierarchical clustering
}
\description{
Calculate a distance matrix for use with hclust
}
