% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOrthologSummary.R
\name{GetOrthologSummary}
\alias{GetOrthologSummary}
\title{Summarize statistics about ortholog predictions.}
\usage{
GetOrthologSummary(OrthologsObject, GeneCalls, DBPath,
  SimilarityScores = FALSE, Verbose = FALSE)
}
\arguments{
\item{OrthologsObject}{An object of class "Orthologs"}

\item{GeneCalls}{A list of dataframes, one for each genome in SyntenyObject, with named columns "Index", "Start", "Stop", "Strand", and an optional "Annotation" column.}

\item{DBPath}{A character vector specifying a sqlite database built by DECIPHER}

\item{SimilarityScores}{A logical indicating whether to get the global alignment scores for all predicted orthologs}

\item{Verbose}{Run with progress bar, return total upon completion.}
}
\description{
Summarize statistics about ortholog predictions.
}
\examples{
GetOrthologSummary()
}
\keyword{Alignment}
\keyword{Gene}
\keyword{Length,}
\keyword{Orthology,}
\keyword{Synteny,}
