% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{getOverlapClusters}
\alias{getOverlapClusters}
\title{getOverlapClusters}
\usage{
getOverlapClusters(sample_probes)
}
\arguments{
\item{sample_probes}{logical matrix of probe calls, where rows are probe ids
and columns are samples}
}
\value{
data.frame
}
\description{
This function finds all clusters where at least one-probe is called
in a sample consecutively across the tiled protein
A way of filtering the clusters to concentrate on areas where there is
significant signal.
}
\examples{
mat = matrix(runif(25) >= 0.5, nrow=5)
rownames(mat) = paste0("A;",seq_len(nrow(mat)))
getOverlapClusters(mat)
}
