% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getProteinStart}
\alias{getProteinStart}
\title{Get the amino-acid starting position of the probe within the protein.}
\usage{
getProteinStart(probes)
}
\arguments{
\item{probes}{vector of probes (i.e. c("A;1", "A;2"))}
}
\value{
starting locations of the probes with their associated proteins
}
\description{
Get the amino-acid starting position of the probe within the protein.
}
\examples{
getProteinStart("A;1")
getProteinStart("B;2")
getProteinStart(c("A;1","B;2"))
}
