% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getProteinTiling}
\alias{getProteinTiling}
\title{Get Protein Tiling}
\usage{
getProteinTiling(probes, return.vector = TRUE)
}
\arguments{
\item{probes}{vector of probes (i.e. A;1, A;2)}

\item{return.vector}{Return result as vector or return as data.frame}
}
\value{
For each protein, the estimating tiling (spacing) of the probes
across the amino acid sequence.
}
\description{
Given a set of probes, estimate the tiling of the probes
across the protein.
}
\examples{
getProteinTiling(c("A;1","A;2","A;3", "B;2","B;3", "C;1", "C;3"))
}
