% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope.R
\name{getSequenceAnnotations}
\alias{getSequenceAnnotations}
\title{Obtain Sequence Annotations for Epitopes}
\usage{
getSequenceAnnotations(epitopes, probe_meta, debug = FALSE)
}
\arguments{
\item{epitopes}{vector of epitope ids}

\item{probe_meta}{data.frame with the PROBE_ID, SEQ_ID, and PROBE_SEQUENCE
columns defined}

\item{debug}{output debugging information}
}
\value{
data.frame in same order of the epitopes parameter with sequence
annotations
}
\description{
Obtain Sequence Annotations for Epitopes
}
