% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getSequenceMatToProbeMat}
\alias{getSequenceMatToProbeMat}
\title{Create a matrix for sequence to probes}
\usage{
getSequenceMatToProbeMat(probe_meta, debug = FALSE)
}
\arguments{
\item{probe_meta}{data.frame with PROBE_ID and PROBE_SEQUENCE columns}

\item{debug}{print out debugging information}
}
\value{
a sparseMatrix that would convert a sequence matrix to probe
matrix upon multiplication
}
\description{
Create a matrix for sequence to probes
}
\examples{
data(heffron2020_wuhan)
probe_meta <- attr(heffron2020_wuhan, "probe_meta")
pData <- attr(heffron2020_wuhan, "pData")
probe_mat = convertSequenceMatToProbeMat(heffron2020_wuhan, probe_meta)

}
