% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hamming}
\alias{hamming}
\title{Calculate hamming distance}
\usage{
hamming(X)
}
\arguments{
\item{X}{logical matrix of calls}
}
\value{
matrix of hamming distances
}
\description{
Calculate hamming distance
}
\examples{
mat = matrix(runif(100) >= 0.5, nrow=10)
rownames(mat) = paste0("A;",seq_len(nrow(mat)))
hamming(mat)
}
