% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hamming_dist}
\alias{hamming_dist}
\title{Calculate normalized hamming distance (jaccard?)}
\usage{
hamming_dist(X)
}
\arguments{
\item{X}{logical matrix of calls}
}
\value{
distance object normalized by the number of number of columns
}
\description{
Calculate normalized hamming distance (jaccard?)
}
