% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HERON-package.R
\docType{package}
\name{HERON-package}
\alias{HERON}
\alias{HERON-package}
\title{HERON: Helpers for River Observation}
\description{
Supplementary functions for workflows including the "EGRET" and "SiZer" R packages. Meant to improve reproducibility for common operations following use of these packages.
}
\author{
\strong{Maintainer}: Nicholas Lyon \email{nickjlyon@gmail.com} (\href{https://orcid.org/0000-0003-3905-1078}{ORCID})

Other contributors:
\itemize{
  \item Joanna Carey \email{jcarey@babson.edu} (\href{https://orcid.org/0000-0003-2365-9185}{ORCID}) [contributor]
  \item Kathi Jo Jankowski \email{kjankowski@usgs.gov} (\href{https://orcid.org/0000-0002-3292-4182}{ORCID}) [contributor]
  \item Keira Johnson \email{johnkeir@oregonstate.edu} (\href{https://orcid.org/0000-0003-0671-3901}{ORCID}) [contributor]
  \item Gabe De La Rosa \email{delarosa@nceas.ucsb.edu} [contributor]
}

}
\keyword{internal}
