% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron.R
\name{heron}
\alias{heron}
\title{Aire d'un triangle}
\usage{
heron(a, b, c)
}
\arguments{
\item{a}{La taille du premier coté du triangle}

\item{b}{La taille du deuxième coté du triangle}

\item{c}{La taille du troisième coté du triangle}
}
\value{
Cette fonction retourne l'aire du triangle
}
\description{
Calcul de l'aire d'un triangle selon la formule de Héron
}
\examples{
heron(3,4,5)
Output : 6

heron(4,7,9)
Output : 13,42
}
