% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{heron2}
\alias{heron2}
\title{heron2}
\usage{
heron2(list_triangle)
}
\arguments{
\item{list_triangle}{Une liste de triangles, où chaque triangle est défini par ses coordonnées (x, y).}
}
\value{
La somme des aires des triangles dans la liste de triangles.
}
\description{
Cette fonction calcule la somme des aires de triangles dans une liste de triangles en utilisant la formule d'Heron.
}
