% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron_bis.R
\name{heron_bis}
\alias{heron_bis}
\title{Calcul de l'aire d'un triangle avec la formule de héron}
\usage{
heron_bis(a, b, c)
}
\arguments{
\item{a}{longueur du premier côté du triangle}

\item{b}{longueur du deuxième côté du triangle}

\item{c}{longueur du troisième côté du triagnle}
}
\value{
l'aire du triangle
}
\description{
Calcul de l'aire d'un triangle avec la formule de héron
}
\examples{
heron_bis(6, 4, 3)
heron_bis(7, 4, 3)
}
