% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HERONEpitopeDataSet-class}
\alias{HERONEpitopeDataSet-class}
\alias{.HERONEpitopeDataSet}
\alias{HERONEpitopeDataSet}
\title{HERONEpitopeDataSet object and constructors}
\usage{
HERONEpitopeDataSet(pvalue, ...)
}
\arguments{
\item{pvalue}{calculate epitope p-value matrix}
}
\description{
TODO
}
