% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HERONProbeDataSet-class}
\alias{HERONProbeDataSet-class}
\alias{.HERONProbeDataSet}
\alias{HERONProbeDataSet}
\title{HERONProbeDataSet object and constructors}
\usage{
HERONProbeDataSet(...)
}
\arguments{
\item{...}{arguments provided to \code{SummarizedExperiment}, including
metadata.}
}
\description{
\code{HERONProbeDataSet} is a subclass of \code{RangedSummarizedExperiment}
used to hold assay information on the probe level
}
\examples{
pds <- HERONProbeDataSet()

}
