% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HERONProteinDataSet-class}
\alias{HERONProteinDataSet-class}
\alias{.HERONProteinDataSet}
\alias{HERONProteinDataSet}
\title{HERONProteinDataSet object and constructors}
\usage{
HERONProteinDataSet(pvalue, ...)
}
\arguments{
\item{pvalue}{calculated protein p-value matrix}

\item{...}{arguments provided to \code{SummarizedExperiment}, including
metadata}
}
\description{
\code{HERONProteinDataSet} is a subclass of \code{SummarizedExperiment}
used to hold assay information on the protein-level
}
