% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_day.R
\name{hydro_day}
\alias{hydro_day}
\title{Convert from Calendar Years to Hydro Years}
\usage{
hydro_day(cal_date = NULL, start_month = 10)
}
\arguments{
\item{cal_date}{(date) calendar date to convert}

\item{start_month}{(numeric) number of month to start the hydro year}
}
\description{
Hydrological years begin in October so this function converts calendar dates into hydrologic dates.
}
