% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_inflections.R
\name{id_inflections}
\alias{id_inflections}
\title{Identify Inflection Points in Slope}
\usage{
id_inflections(raw_data = NULL, sizer_data = NULL, x = NULL, y = NULL)
}
\arguments{
\item{raw_data}{(dataframe) data object}

\item{sizer_data}{(dataframe) object returned by `sizer_slice` or `sizer_aggregate`}

\item{x}{(character) column name in `raw_data` that is the x-axis}

\item{y}{(character) column name in `raw_data` that is the y-axis}
}
\description{
"Breaks" a given continuous variable at the *inflection points* identified by `sizer_aggregate` or `sizer_slice` and creates a column (named `groups`) that contains these group designations. Additionally, it identifies the start and end value of X for each group in two other new columns (named `start` and `end`, respectively)
}
