% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_slope_changes.R
\name{id_slope_changes}
\alias{id_slope_changes}
\title{Identify Changes in Slope}
\usage{
id_slope_changes(
  raw_data = NULL,
  x = NULL,
  y = NULL,
  sizer_data = NULL,
  group_dig = 6
)
}
\arguments{
\item{raw_data}{(dataframe) data object}

\item{x}{(character) column name in `raw_data` that is the x-axis}

\item{y}{(character) column name in `raw_data` that is the y-axis}

\item{sizer_data}{(dataframe) object returned by `sizer_slice` or `sizer_aggregate`}

\item{group_dig}{(number) digits to round group edges to}
}
\description{
"Breaks" a given continuous variable at the *slope changes* identified by `sizer_aggregate` or `sizer_slice` and creates a column (named `groups`) that contains these group designations. Additionally, it identifies the start and end value of X for each group in two other new columns (named `start` and `end`, respectively)
}
