% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosts_fast.R
\name{kostPolyFitFast}
\alias{kostPolyFitFast}
\title{Calculate Kost's approximation of covariance}
\usage{
kostPolyFitFast(cor_in)
}
\arguments{
\item{cor_in}{correlation between two n x n data vectors.}
}
\value{
Kost's approximation of the covariance between the -log cumulative distributions. This is calculated with a cubic polynomial fit.
}
\description{
Calculate Kost's approximation of covariance
}
