% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{log2Transform}
\alias{log2Transform}
\title{log2 transform the "exprs" assay}
\usage{
log2Transform(se)
}
\arguments{
\item{se}{SummarizedExperiment with "exprs" assay}
}
\value{
SummarizedExperiment with "exprs" assay log2 transformed
}
\description{
log2 transform the "exprs" assay
}
\examples{
data(heffron2021_wuhan)
assay(heffron2021_wuhan, "exprs") <- 2^assay(heffron2021_wuhan, "exprs")
res <- log2Transform(heffron2021_wuhan)
}
