% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{makeCalls}
\alias{makeCalls}
\title{Title}
\usage{
makeCalls(padj_mat, padj_cutoff = 0.05)
}
\arguments{
\item{padj_mat}{adjusted p-value matrix}

\item{padj_cutoff}{adjusted cutoff}
}
\value{
list of results
sample => calls made on the sample (column)-level
k_of_n => data.frame with K of N statistics.
}
\description{
Title
}
