% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{makeProteinCalls}
\alias{makeProteinCalls}
\title{Make Protein-level Calls}
\usage{
makeProteinCalls(prot_ds, padj_cutoff = 0.05)
}
\arguments{
\item{prot_ds}{HERONProteinDataSet with the "padj" assay}

\item{padj_cutoff}{cutoff to use}
}
\value{
HERONProteinDataSet with the "calls" assay added
}
\description{
Make Protein-level Calls
}
\examples{
data(heffron2021_wuhan)
probe_meta <- attr(heffron2021_wuhan, "probe_meta")
seq_pval_res <- calcCombPValues(heffron2021_wuhan)
pr_pval_res <- convertSequenceDSToProbeDS(seq_pval_res, probe_meta)
pr_calls_res <- makeProbeCallsPDS(pr_pval_res)
epi_segments_uniq_res <- findEpitopeSegmentsPDS(
    PDS_obj = pr_calls_res,
    segment_method = "unique"
)
epi_padj_uniq <- calcEpitopePValuesProbeDS(
    probe_pds = pr_calls_res,
    epitope_ids = epi_segments_uniq_res,
    metap_method = "wilkinsons_max1"
)
prot_padj_uniq <- calcProteinPValuesEpitopeDS(
    epitope_ds = epi_padj_uniq,
    metap_method = "tippets"
)
prot_calls <- makeProteinCalls(prot_padj_uniq)
}
