% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{min_max}
\alias{min_max}
\title{Cap vector at minimum/maximum values}
\usage{
min_max(val, min.value, max.value)
}
\arguments{
\item{val}{vector of values to cap}

\item{min.value}{minimum value}

\item{max.value}{maximum value}
}
\value{
vector of capped values
}
\description{
Cap vector at minimum/maximum values
}
\examples{
min_max(10, 1, 5)
}
