% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBIGFFParser.R
\name{NCBIGFFParser}
\alias{NCBIGFFParser}
\title{A function for parsing genomic General Format Files directly from the NCBI ftp server}
\usage{
NCBIGFFParser(GFFAddress, Verbose = FALSE)
}
\arguments{
\item{GFFAddress}{a vector of characters for ftp addresses of gff.gz files}
}
\description{
This parser is specifically built for gff files that have been produced by NCBI's annotation pipeline
}
\examples{
GFFParser()
}
\keyword{GFF}
