% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NucleotideOverlap.R
\name{NucleotideOverlap}
\alias{NucleotideOverlap}
\title{A function for determining nucleotide overlap that links genes.}
\usage{
NucleotideOverlap(SyntenyObject, GeneCalls, LimitIndex = TRUE,
  OutputFormat = "Normal", Verbose = FALSE)
}
\arguments{
\item{SyntenyObject}{An object of class "Synteny".}

\item{GeneCalls}{A named list of data.frames, one for each genome in SyntenyObject, with named columns "Index", "Start", "Stop", "Strand", and an optional "Annotation" column.}

\item{LimitIndex}{Use only the first index of all selected genomes, defaults to TRUE.}

\item{OutputFormat}{"Comprehensive", "Normal", or "Sparse" for how much data to return.}

\item{Verbose}{Run with progress bar, return total upon completion.}

}
\description{
A function for determining nucleotide overlap that links genes.
}
\examples{
NucleotideOverlap()
}
\keyword{Orthology}
\keyword{Synteny}
