% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{oneHitEpitopes}
\alias{oneHitEpitopes}
\title{Find One-hit epitopes}
\usage{
oneHitEpitopes(sample_epitopes)
}
\arguments{
\item{sample_epitopes}{logical epitope matrix from makeCalls}
}
\value{
vector of one-hit, one-probe epitopes
}
\description{
Find One-hit epitopes
}
\examples{
hit_mat = data.frame(
row.names = c("A_1_1","A_2_2","A_3_3","A_4_4"),
sample1 = c(TRUE, FALSE, FALSE, TRUE),
sample2 = c(TRUE, TRUE, FALSE, FALSE),
sample3 = c(TRUE, TRUE, FALSE, FALSE)
)
oneHitEpitopes(hit_mat)
}
