% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{oneHitProbes}
\alias{oneHitProbes}
\title{Find one hit probes}
\usage{
oneHitProbes(sample_probes)
}
\arguments{
\item{sample_probes}{logical probe matrix from makeCalls}
}
\value{
vector of probes that are one-hits
}
\description{
Find one hit probes
}
\examples{
hit_mat = data.frame(
row.names = c("A;1","A;2","A;3","A;4"),
sample1 = c(TRUE, FALSE, FALSE, TRUE),
sample2 = c(TRUE, TRUE, FALSE, FALSE),
sample3 = c(TRUE, TRUE, FALSE, FALSE)
)
oneHitProbes(hit_mat)
}
