% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{p_adjust_ds}
\alias{p_adjust_ds}
\title{Adjust an assay of p-values column-by-column}
\usage{
p_adjust_ds(obj, method = "BH")
}
\arguments{
\item{obj}{SummarizedExperiment with a "pvalue" assay}

\item{method}{what adjustment algorithm to use (see p.adjust)}
}
\description{
Adjust an assay of p-values column-by-column
}
