% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{p_adjust_mat}
\alias{p_adjust_mat}
\title{Adjust a matrix of p-values column-by-column}
\usage{
p_adjust_mat(pvalues_mat, method = "BH")
}
\arguments{
\item{pvalues_mat}{matrix of p-values}

\item{method}{what adjustment algorithm to use (see p.adjust)}
}
\value{
matrix of column-by-column adjusted p-values
}
\description{
Adjust a matrix of p-values column-by-column
}
\examples{
mat = matrix(runif(25) >= 0.5, nrow=5)
rownames(mat) = paste0("A;",seq_len(nrow(mat)))
p_adjust_mat(mat)
}
