% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{probeHitSupported}
\alias{probeHitSupported}
\title{Find probe hits that are supported by a consecutive probe or another sample}
\usage{
probeHitSupported(
  hit_mat,
  probes = rownames(hit_mat),
  proteins = getProteinLabel(probes),
  positions = getProteinStart(probes),
  tiling = getProteinTiling(probes)
)
}
\arguments{
\item{tiling}{}
}
\value{

}
\description{
Find probe hits that are supported by a consecutive probe or another sample
}
