% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pvalue_to_zscore}
\alias{pvalue_to_zscore}
\title{Convert p-value matrix to a z-score matrix}
\usage{
pvalue_to_zscore(m, one.sided = TRUE, log.p = FALSE, inf_zscore = 16)
}
\arguments{
\item{one.sided}{p-values one-sided}

\item{log.p}{are p-values log transformed?}

\item{inf_zscore}{infinite z-scores are capped to this value}

\item{mat.in}{matrix of p-values}
}
\value{

}
\description{
Convert p-value matrix to a z-score matrix
}
