% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{quantileNormalize}
\alias{quantileNormalize}
\title{Normalize a matrix/data.frame using quantile normaliztion}
\usage{
quantileNormalize(in_mat)
}
\arguments{
\item{in_mat}{matrix or data.frame of numeric values to be normalized}
}
\value{
normalized matrix
}
\description{
Normalize a matrix/data.frame using quantile normaliztion
}
