% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadMCL.R
\name{ReadMCL}
\alias{ReadMCL}
\title{Read an MCL output file into R.}
\usage{
ReadMCL(FileName, Verbose = FALSE)
}
\arguments{
\item{FileName}{A character vector specifying the location of an MCL file to read.}
}
\description{
Read an MCL output file into R.
}
\examples{
ReadMCL()
}
\keyword{MCL,}
\keyword{Orthology}
