% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResolveConflicts.R
\name{ResolveConflicts}
\alias{ResolveConflicts}
\title{A function for removing conflicting ortholog predictions}
\usage{
ResolveConflicts(SummaryObject, ResolveBy = "TotalCoverage", Verbose = FALSE)
}
\arguments{
\item{OrhtologSummary}{A data.frame containing metrics about predicted ortholog pairs. The rownames of this data.frame are the labels of the pair of genes predicted to be orthologs.}

\item{ResolveBy}{Which attribute to resolve conflicting predictions by. Resolving by Coverage or Similarity will result in keeping the prediction providing the higher score in either metric. Resolving by NormGeneDiff, NormDeltaStart, or NormDeltaStop will result in keeping the prediction with the minimum value of that metric.}
}
\description{
Syntenic hits that exist in regions of gene overlap result in putative predictions of orthology from a single gene in one genome to two genes in a second genome. One of these predictions is presumably the "intended" prediction, while the other is spurious. Other genomic phenomena result in predictions that are likewise spurious and must be resolved. ResolveConflicts removes these spurious predictions based on the metric of choice for the user.
}
\examples{
ResolveConflicts()
}
\keyword{Orthologs}
