% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizer_lm.R
\name{sizer_lm}
\alias{sizer_lm}
\title{Linear Models with SiZer}
\usage{
sizer_lm(data = NULL, x = NULL, y = NULL, group_col = NULL)
}
\arguments{
\item{data}{(dataframe) data object returned by `id_inflections` or `id_slope_changes`}

\item{x}{(character) column name in `data` that is the x-axis}

\item{y}{(character) column name in `data` that is the y-axis}

\item{group_col}{(character) column name in `data` that identifies sections of trendline that share a common slope (see object returned by `id_inflections` or `id_slope_changes`)}
}
\value{
(list) containing (1) the model summary statistics and (2) the the estimates of the intercept and line
}
\description{
Fits separate linear models for each group of the line between the break points identified by `id_inflections` or `id_slope_changes`
}
