% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizer_plot.R
\name{sizer_plot}
\alias{sizer_plot}
\title{Create SiZer Base R Plot}
\usage{
sizer_plot(sizer_object = NULL, bandwidth_vec = c(3, 6, 9))
}
\arguments{
\item{sizer_object}{(SiZer) object returned by `SiZer::SiZer`}

\item{bandwidth_vec}{(numeric) bandwidths to apply a horizontal line in the plot}
}
\description{
Creates the base R `plot` returned for `SiZer` objects and adds horizontal lines at three bandwidths
}
