% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizer_slice.R
\name{sizer_slice}
\alias{sizer_slice}
\title{Identify Inflection Points for One SiZer Bandwidth}
\usage{
sizer_slice(sizer_object = NULL, bandwidth = NULL)
}
\arguments{
\item{sizer_object}{(SiZer) object returned by `SiZer::SiZer`}

\item{bandwidth}{(numeric) bandwidth within `sizer_object` to extract}
}
\value{
(dataframe) slope type (flat, positive, or negative) identified by `SiZer::SiZer` at each x value for the specified bandwidth
}
\description{
Identifies the inflection points (i.e., +/- or -/+ slope changes) for the *one* specified bandwidth. Note that this bandwidth need not be an exact match for what is given to the `sizer_object` argument but should be close
}
