% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{smoothProbeDS}
\alias{smoothProbeDS}
\title{Smooth probes across protein tiling}
\usage{
smoothProbeDS(probe_ds, w = 2, eps = 1e-06)
}
\arguments{
\item{probe_ds}{HERONProbeDataSet to smooth}

\item{w}{smoothing width, probes +/- w/2 before and after are used for smoothing}

\item{eps}{error tolerance}
}
\value{
HERONProbeDataSet with smoothed data in exprs object
}
\description{
Smooth probes across protein tiling
}
\examples{
data(heffron2021_wuhan)
probe_meta <- attr(heffron2021_wuhan, "probe_meta")
probe_ds <- convertSequenceDSToProbeDS(heffron2021_wuhan, probe_meta)
smoothed_ds <- smoothProbeDS(probe_ds)
}
