% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{smoothProbeMat}
\alias{smoothProbeMat}
\title{Smooth probes across protein tiling}
\usage{
smoothProbeMat(probe_mat, w = 2, eps = 1e-06)
}
\arguments{
\item{probe_mat}{probe matrix}

\item{w}{smoothing width, probes w/2 before and after are used for smoothing}

\item{eps}{error tolerance}
}
\value{
probe matrix with smoothed data
}
\description{
Smooth probes across protein tiling
}
\examples{
data(heffron2021_wuhan)
probe_meta <- attr(heffron2021_wuhan, "probe_meta")
probe_mat <- convertSequenceMatToProbeMat(heffron2021_wuhan, probe_meta)
smoothed_mat <- smoothProbeMat(probe_mat)
}
