% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubCatalog.R
\name{SubCatalog}
\alias{SubCatalog}
\title{A function for doing simple network analysis on sets of homologs}
\usage{
SubCatalog(EdgeMatrix, SubDivision, Verbose = FALSE)
}
\arguments{
\item{EdgeMatrix}{A matrix of predicted orthologs, with each row being an "edge" consisting of two nodes, each being genes in a predicted orthologous pair. The nodes are located in a column that specifies a genome.}

\item{SubDivision}{Selected rows to build an edge network for}
}
\description{
The function `Catalog` is not parallelizable, this is a parallelizable version `Catalog`, but requires an EdgeMatrix instead of an Orthology Object. Lists of these subcatalogs can then be stitched together.
}
\examples{
SubCatalog()
}
\keyword{Homology}
