% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcCombPValues}
\alias{calcCombPValues}
\title{Calculate p-values using the "exprs" assay from the sequence or probe dataset}
\usage{
calcCombPValues(
  obj,
  t.sd_shift = NA,
  t.abs_shift = NA,
  t.paired = FALSE,
  z.sdshift = 0,
  use = "both",
  p.adjust.method = "BH"
)
}
\arguments{
\item{obj}{HERONSequenceDataSet or HERONProbeDataSet}

\item{p.adjust.method}{}
}
\value{
HERONSequenceDataSet/HERONProbeDataSet with the pvalue assay added
}
\description{
Calculate p-values using the "exprs" assay from the sequence or probe dataset
}
