% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcCombPValues}
\alias{calcCombPValues}
\title{Title}
\usage{
calcCombPValues(
  obj,
  t.sd_shift = NA,
  t.abs_shift = NA,
  t.paired = FALSE,
  z.sdshift = 0,
  use = "both",
  p.adjust.method = "BH"
)
}
\arguments{
\item{p.adjust.method}{}
}
\description{
Title
}
