% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcEpitopePValuesMat}
\alias{calcEpitopePValuesMat}
\title{Calculate epitope-level p-values}
\usage{
calcEpitopePValuesMat(
    probe_pvalues,
    epitope_ids,
    metap_method = "wmax1",
    p_adjust_method = "BH"
)
}
\arguments{
\item{probe_pvalues}{matrix of probe p-values}

\item{epitope_ids}{vector of epitope ids}

\item{metap_method}{meta p-value method to use}

\item{p_adjust_method}{what p.adjust method to use.}
}
\value{
matrix of epitope-level p-values
}
\description{
Calculate epitope-level p-values
}
\examples{
data(heffron2020_wuhan)
probe_meta <- attr(heffron2020_wuhan, "probe_meta")
pData <- attr(heffron2020_wuhan, "pData")
pval_res <- calcProbePValuesSeqMat(heffron2020_wuhan, probe_meta, pData)
calls_res <- makeProbeCalls(pval_res)
segments_res <- findEpitopeSegments(probe_calls = calls_res);
epval_res <- calcEpitopePValuesMat(attr(pval_res, "pvalue"), segments_res)
}
