% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{calcMetaPValuesVec}
\alias{calcMetaPValuesVec}
\title{Calculate meta p-values given a vector of lower-level p-values}
\usage{
calcMetaPValuesVec(pvalues, by_list, method = "min_bonf", do.sort = FALSE)
}
\arguments{
\item{pvalues}{vector of p-values}

\item{by_list}{list of groupings for meta-pvalue calculation}

\item{method}{meta p-value method to use}

\item{do.sort}{sort in increasing order?}
}
\value{
vector of meta p-values
}
\description{
Calculate meta p-values given a vector of lower-level p-values
}
