% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcMinFDR}
\alias{calcMinFDR}
\title{Calculate minimum FDRs across samples}
\usage{
calcMinFDR(fdrs, additional_stats = TRUE, sort = TRUE)
}
\arguments{
\item{fdrs}{matrix of adjusted p-values}

\item{additional_stats}{indicator to calculate mean and median FDR}

\item{sort}{sort the resultant matrix by the last N column (increasing)}
}
\value{
matrix of FDR for each K level in n1..N column names
}
\description{
This code calculates the minimum FDR threshold needed
to call K out of N samples, where samples are in the columns
When additional stats is selected, also calculates the mean and median
}
