% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProbePValuesProbeMat}
\alias{calcProbePValuesProbeMat}
\title{Calculate Probe-level p-values}
\usage{
calcProbePValuesProbeMat(
  probe_mat,
  pData,
  t.sd_shift = NA,
  t.abs_shift = NA,
  t.paired = FALSE,
  z.sdshift = 0,
  use = "both",
  p.adjust.method = "BH"
)
}
\arguments{
\item{probe_mat}{numeric matrix, rows as seqs and columns as samples}

\item{pData}{experiment design data.frame}

\item{t.sd_shift}{sd shift multiplier for diff t-test}

\item{t.abs_shift}{abs shift to use for the diff t-test}

\item{t.paired}{do paired t-test}

\item{z.sdshift}{sd shift multiplier for global z-test}

\item{use}{which p-value method(s) to use}

\item{p.adjust.method}{method for adjusting p-values}
}
\value{
matrix of calculated p-values
}
\description{
calculates p-values on the matrix (can be sequence as well), using either a
z-score global, t-stat differential, or a combination of both.
input a matrix that has samples on the columns and sequence probes on the
rows.  pData is a design matrix that describes the layout of the
experiment, similar to the pepStat matrix.
}
\examples{
data(heffron2020_wuhan)
probe_meta <- attr(heffron2020_wuhan, "probe_meta")
pData <- attr(heffron2020_wuhan, "pData")
probe_mat = convertSequenceMatToProbeMat(heffron2020_wuhan, probe_meta)
pval_res <- calcProbePValuesProbeMat(probe_mat, pData)
}
