% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProbePValuesProbeMat}
\alias{calcProbePValuesProbeMat}
\title{Calculate Probe-level p-values}
\usage{
calcProbePValuesProbeMat(
  probe_mat,
  pData,
  t.sd_shift = NA,
  t.abs_shift = NA,
  z.sdshift = 0,
  use = "both",
  p.adjust.method = "BH"
)
}
\arguments{
\item{probe_mat}{matrix of numeric values that where the rows are features
and the columns are samples. assumed to be log-transformed data}

\item{pData}{experiment design data.frame}

\item{t.sd_shift}{shift to use when calculating the differential t-test
p-values using multiples of the calculated standard deviation of the values.
Either t.sd_shift or t.abs_shift should be set}

\item{t.abs_shift}{shift to use when calculating the differential t-test
p-values using an absolute shift. Either t.sd_shift or t.abs_shift should be
set}

\item{z.sdshift}{shift to use when calculating the global z-test p-values
using multiples of the calculated standard deviation across all values in the
matrix.}

\item{use}{what p-value method to use, t - differential t-test, z - global
z-test, both - both the differential t-test and global z-test.}

\item{p.adjust.method}{method for adjusting p-values (multiple testing)}
}
\value{
matrix of p-values calculating on the matrix values and supplied
parameters
}
\description{
calculates p-values on the matrix (can be sequence as well), using either a
z-score global, t-stat differential, or a combination of both.
input a matrix that has samples on the columns and sequence probes on the
rows.  pData is a design matrix that describes the layout of the
experiment, similar to the pepStat matrix.
}
