% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProbePValuesSeqMat}
\alias{calcProbePValuesSeqMat}
\title{Title}
\usage{
calcProbePValuesSeqMat(
  seq_mat,
  probe_meta,
  pData,
  t.sd_shift = NA,
  t.abs_shift = NA,
  z.sdshift = 0,
  use = "both",
  make.plots = TRUE,
  combine = "max"
)
}
\arguments{
\item{combine}{}
}
\description{
Title
}
