% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProbePValuesSeqMat}
\alias{calcProbePValuesSeqMat}
\title{Calculate probe p-vlaues using a sequence matrix.}
\usage{
calcProbePValuesSeqMat(
  seq_mat,
  probe_meta,
  pData,
  t.sd_shift = NA,
  t.abs_shift = NA,
  z.sdshift = 0,
  use = "both",
  make.plots = TRUE,
  combine = "max",
  p.adjust.method = "BH"
)
}
\arguments{
\item{seq_mat}{matrix of values where the rows are sequence identifiers
and the columns are samples}

\item{probe_meta}{data.frame with the columns PROBE_ID and PROBE_SEQUENCE}

\item{pData}{design matrix}

\item{t.sd_shift}{standard deviation shift for differential test}

\item{t.abs_shift}{absolute shift for differential test}

\item{z.sdshift}{standard deviation shift for global test}

\item{use}{use global-test ("z"), differential-test ("t"), or both ("both")}

\item{make.plots}{indicator to make plots of results}

\item{combine}{meta p-value method to use when combining tests.}

\item{p.adjust.method}{p-value adjustment method to use (default BH)}
}
\value{
matrix of adjusted p-values with additional attributes.
}
\description{
Calculate probe p-vlaues using a sequence matrix.
}
