% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProbePValuesTUnpaired}
\alias{calcProbePValuesTUnpaired}
\title{calculate Probe p-values using a Differential unpaired t-test}
\usage{
calcProbePValuesTUnpaired(
  probe_mat,
  pData,
  sd_shift = NA,
  abs_shift = NA,
  debug = FALSE
)
}
\arguments{
\item{probe_mat}{numeric matrix or data.frame of values}

\item{pData}{design data.frame}

\item{sd_shift}{standard deviation shift to use when calculating p-values
Either sd_shift or abs_shift should be set}

\item{abs_shift}{absolute shift to use when calculating p-values}

\item{debug}{output debugging information}
}
\value{
matrix of p-values on the post columns defined in the pData matrix
}
\description{
calculate Probe p-values using a Differential unpaired t-test
}
