% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProbePValuesTUnpaired}
\alias{calcProbePValuesTUnpaired}
\title{Calculate Probe p-values using a differential unpaired t-test}
\usage{
calcProbePValuesTUnpaired(
  probe_mat,
  pData,
  sd_shift = NA,
  abs_shift = NA,
  keep_all_cols = FALSE
)
}
\arguments{
\item{probe_mat}{numeric matrix or data.frame of values}

\item{pData}{design data.frame}

\item{sd_shift}{standard deviation shift to use when calculating p-values
Either sd_shift or abs_shift should be set}

\item{abs_shift}{absolute shift to use when calculating p-values}
}
\value{
matrix of p-values on the post columns defined in the pData matrix
}
\description{
Calculate Probe p-values using a differential unpaired t-test
}
\examples{
data(heffron2021_wuhan)
probe_meta <- attr(heffron2021_wuhan, "probe_meta")
pData <- attr(heffron2021_wuhan, "pData")
pval_res <- calcProbePValuesTUnpaired(heffron2021_wuhan, pData)
}
