% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calcProteinPValuesMat}
\alias{calcProteinPValuesMat}
\title{Calculate protein-level p-values}
\usage{
calcProteinPValuesMat(
  epitope_pvalues_mat,
  metap_method = "wmin1",
  p_adjust_method = "BH"
)
}
\arguments{
\item{epitope_pvalues_mat}{matrix of epitope-level p-values}

\item{metap_method}{meta p-value method to use}

\item{p_adjust_method}{p.adjust method to use}
}
\value{
matrix of protein-level p-values
}
\description{
Calculate protein-level p-values
}
\examples{
data(heffron2021_wuhan)
probe_meta <- attr(heffron2021_wuhan, "probe_meta")
pData <- attr(heffron2021_wuhan, "pData")
pval_res <- calcProbePValuesSeqMat(heffron2021_wuhan, probe_meta, pData)
calls_res <- makeProbeCalls(pval_res)
segments_res <- findEpitopeSegments(probe_calls = calls_res);
epval_res <- calcEpitopePValuesMat(pval_res, segments_res)
ppval_res <- calcProteinPValuesMat(epval_res)
}
