% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{catSequences}
\alias{catSequences}
\title{Concatenate sequences together based upon their start positions.
Assumes the probe sequences have an overlap.}
\usage{
catSequences(positions, sequences)
}
\arguments{
\item{positions}{start positions of probes in protein}

\item{sequences}{probe sequences of probes}
}
\value{
concatenated sequence (character)
}
\description{
Concatenate sequences together based upon their start positions.
Assumes the probe sequences have an overlap.
}
\examples{
positions <- c(1,2)
sequences <- c("MSGSASFEGGVFSPYL", "SGSASFEGGVFSPYLT")
catSequences(positions, sequences)
}
