% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convertSequenceDSToProbeDS}
\alias{convertSequenceDSToProbeDS}
\title{Convert HERONSequenceDataSet to HERONProbeDataSet}
\usage{
convertSequenceDSToProbeDS(seq_ds, probe_meta)
}
\arguments{
\item{seq_ds}{a HERONSequenceDataSet object}

\item{probe_meta}{data.frame with the PROBE_SEQUENCE, PROBE_ID columns}
}
\value{
HERONProbeDataSet
}
\description{
Convert HERONSequenceDataSet to HERONProbeDataSet
}
