% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{findBlocksT}
\alias{findBlocksT}
\title{Find consecutive probes}
\usage{
findBlocksT(prot_df, protein_tiling)
}
\arguments{
\item{prot_df}{data.frame with the Protein and Starting position of the
probe}

\item{protein_tiling}{tiling for information for each protein}
}
\value{
data.frame with the Protein, Start, Stop, and Number.Of.Probes
columns
}
\description{
Find consecutive probes
}
\examples{
probes = c("A;1","A;2","A;3", "A;5","A;6", "A;8")
prot_df = data.frame(
    Protein = getProteinLabel(probes),
    Pos = getProteinStart(probes)
)
findBlocksT(prot_df)
}
