% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{findEpitopeSegmentsUnique}
\alias{findEpitopeSegmentsUnique}
\title{Title}
\usage{
findEpitopeSegmentsUnique(probe_calls_res)
}
\arguments{
\item{probe_calls_res}{Results from makeProbeCalls}
}
\description{
Title
}
