% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{getClusterSegmentsHClust}
\alias{getClusterSegmentsHClust}
\title{Find Cluster Segments using hclust}
\usage{
getClusterSegmentsHClust(
  sample_probes_sub,
  cluster_id,
  do.plot = FALSE,
  cutoff = "silhouette",
  dist.method = "hamming",
  dist_mat2 = getDistMat2(sample_probes_sub, dist.method = dist.method),
  hc = stats::hclust(stats::as.dist(dist_mat2), method = "complete")
)
}
\arguments{
\item{sample_probes_sub}{logical or numerical (z-score) matrix of probe calls}

\item{cluster_id}{epitope identifier of cluster to further segment}

\item{do.plot}{make plots of result}

\item{cutoff}{cutoff to use when finding the number of clusters}

\item{dist.method}{distance algorithm to use}

\item{dist_mat2}{calculated using getDistMat2 (cached result)}

\item{hc}{calculated using hclust (cached result)}
}
\value{
vector of segmented epitope identifiers
}
\description{
Find Cluster Segments using hclust
}
