% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{getClusterSegmentsSkater}
\alias{getClusterSegmentsSkater}
\title{Title}
\usage{
getClusterSegmentsSkater(
  sample_probes_sub,
  cluster_id,
  dist.method = "euclidean",
  cutoff = "silhouette",
  debug = FALSE
)
}
\arguments{
\item{sample_probes_sub}{TODO}

\item{cluster_id}{TODO}

\item{dist.method}{TODO}

\item{cutoff}{TODO}

\item{debug}{TODO}
}
\value{
vector of segments
}
\description{
Title
}
