% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{getClusterSegmentsSkater}
\alias{getClusterSegmentsSkater}
\title{Get Segmentation Using Skater}
\usage{
getClusterSegmentsSkater(
    sample_probes_sub,
    cluster_id,
    dist_method = "euclidean",
    cutoff = "silhouette"
)
}
\arguments{
\item{sample_probes_sub}{matrix of probe calls}

\item{cluster_id}{cluster identification}

\item{dist_method}{distance method to use}

\item{cutoff}{cutoff to use, silhouette supported.}
}
\value{
vector of segments
}
\description{
Acceptable dist.methods are "euclidean", "hamming",
"maximum", "manhattan", "canberra", "binary", "minkowski"
Minkowski is hardcoded to have a power (p) of number of columns
in sample_probes_sub.
}
