% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope.R
\name{getEpitopeIDsToProbeIDs}
\alias{getEpitopeIDsToProbeIDs}
\title{Get probe ids from a vector of epitope ids}
\usage{
getEpitopeIDsToProbeIDs(epitope_ids, tiling = 1)
}
\arguments{
\item{epitope_ids}{vector of epitope identifiers}

\item{tiling}{tling of probes across proteins}
}
\value{
data.frame of epitope_to_probe mappings
}
\description{
Get probe ids from a vector of epitope ids
}
\examples{
getEpitopeIDsToProbeIDs(c("A_1_5","C_8_12"))
}
