% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope.R
\name{getEpitopeProtein}
\alias{getEpitopeProtein}
\title{Obtain Protein Id from Epitope ID}
\usage{
getEpitopeProtein(epitope_ids)
}
\arguments{
\item{epitope_ids}{vector of epitope identifier character strings}
}
\value{
vector of protein labels
}
\description{
Format of EpitopeID is A_B_C, where A is the protein label
B is the protein start position of the first probe in the epitope and
C is the protein start position of the last probe in the epitope.
}
\examples{
getEpitopeProtein("Prot1_1_5")
}
