% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{getHClustDistMat}
\alias{getHClustDistMat}
\title{Calculate a distance matrix for use with hclust}
\usage{
getHClustDistMat(sample_probes_sub, dist.method, debug = FALSE)
}
\arguments{
\item{sample_probes_sub}{TODO}

\item{dist.method}{TODO}

\item{debug}{TODO}
}
\value{
matrix that can be converted to a distance for use with complete
hierarchical clustering
}
\description{
Calculate a distance matrix for use with hclust
}
