% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{getKofN}
\alias{getKofN}
\title{Get K of N statistics from an experiment with padj and calls}
\usage{
getKofN(obj)
}
\arguments{
\item{obj}{}
}
\description{
Get K of N statistics from an experiment with padj and calls
}
