% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_finding.R
\name{getOverlapClusters}
\alias{getOverlapClusters}
\title{getOverlapClusters}
\usage{
getOverlapClusters(
  sample_probes,
  protein_tiling = getProteinTiling(rownames(sample_probes))
)
}
\arguments{
\item{sample_probes}{from makeProbeCalls(...), logical matrix of probe calls}

\item{protein_tiling}{tiling for each protein}
}
\value{
data.frame
}
\description{
This function finds all clusters where at least one-probe is called
in a sample consecutively across the tiled protein
A way of filtering the clusters to concentrate on areas where there is
significant signal.
}
