% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getProteinStart}
\alias{getProteinStart}
\title{Get the amino-acid starting position of the probe within the protein.}
\usage{
getProteinStart(probes)
}
\arguments{
\item{probes}{vector of probes (i.e. c("A;1", "A;2"))}
}
\value{

}
\description{
Get the amino-acid starting position of the probe within the protein.
}
