% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getProteinTiling}
\alias{getProteinTiling}
\title{Get Protein Tiling}
\usage{
getProteinTiling(probes, return.vector = TRUE)
}
\arguments{
\item{probes}{vector of probes (i.e. A;1, A;2)}

\item{return.vector}{Return result as vector or return as data.frame}
}
\value{

}
\description{
Given a set of probes, estimate the tiling of the probes
across the protein.
}
