% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope.R
\name{getSequenceAnnotations}
\alias{getSequenceAnnotations}
\title{Obtain Sequence Annotations for Epitopes}
\usage{
getSequenceAnnotations(epitopes, probe_meta)
}
\arguments{
\item{epitopes}{vector of epitope ids}

\item{probe_meta}{data.frame with the PROBE_ID and PROBE_SEQUENCE
columns defined}
}
\value{
data.frame in same order of the epitopes parameter with sequence
annotations
}
\description{
Obtain Sequence Annotations for Epitopes
}
\examples{
probe_meta = data.frame(
PROBE_ID=c("A;1","A;2"),
PROBE_SEQUENCE = c("MSGSASFEGGVFSPYL","SGSASFEGGVFSPYLT"))
getSequenceAnnotations("A_1_2", probe_meta)
}
