% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heffron2020_wuhan}
\alias{heffron2020_wuhan}
\title{SARS CoV-2 Wuhan Peptide Binding Array data}
\format{
## `heffron2020_wuhan`
A data frame with X rows and Y columns:
}
\source{
<https://dholk.primate.wisc.edu/_webdav/dho/sequencing/Polypeptide%20Microarrays/public/COVID_19/%40files>
}
\usage{
heffron2020_wuhan
}
\description{
A subset of data from the paper
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8245122/ publication.
}
\keyword{datasets}
