% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron.R
\name{heron}
\alias{heron}
\title{Calcul de l'aire d'un triangle à partir de la longueur de ses côtés}
\usage{
heron(a, b, c)
}
\arguments{
\item{a}{Longueur du premier côté du triangle.}

\item{b}{Longueur du deuxième côté du triangle.}

\item{c}{Longueur du troisième côté du triangle.}
}
\value{
L'aire du triangle.
}
\description{
heron
}
\examples{
heron(3, 4, 5)
heron(6, 8, 10)
}
