% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{heron}
\alias{heron}
\title{heron}
\usage{
heron(a, b, c)
}
\arguments{
\item{a}{Longueur d'un côté du triangle}

\item{b}{Longueur d'un côté du triangle}

\item{c}{Longueur d'un côté du triangle}
}
\value{
L'aire du triangle
}
\description{
Calcule l'aire d'un triangle à partir de ses côtés
}
\examples{
heron(3, 4, 5)
heron(6, 8, 10)

}
