% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron.R
\name{heron}
\alias{heron}
\title{Heron's Formula}
\usage{
heron(a, b, c)
}
\arguments{
\item{a}{a positive float}

\item{b}{a positive float}

\item{c}{a positive float}
}
\value{
area of a triangle
}
\description{
Heron's Formula
}
\examples{
heron(3,3,3)
heron(2,2,2)

}
