% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron.R
\name{heron}
\alias{heron}
\title{Calcule l'aire d'un triangle à partir de ses côtés}
\usage{
heron(a, b, c)
}
\arguments{
\item{a}{Le premier côté du triangle}

\item{b}{Le deuxième côté du triangle}

\item{c}{Le troisième côté du triangle}
}
\value{
L'aire du triangle
}
\description{
Calcule l'aire d'un triangle à partir de ses côtés
}
\examples{
heron(3,4,5)
heron(3,2,2)
}
