% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron.R
\name{heron}
\alias{heron}
\title{Calcul de l'aire d'un triangle avec la formule de héron}
\usage{
heron(a, b, c)
}
\arguments{
\item{a}{longueur du premier côté du triangle}

\item{b}{longueur du deuxième côté du triangle}

\item{c}{longueur du troisième côté du triagnle}
}
\value{
l'aire du triangle
}
\description{
Calcul de l'aire d'un triangle avec la formule de héron
}
\examples{
heron(6,4,3)
heron(7,5,3)
}
