% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron.R
\name{heron}
\alias{heron}
\title{heron}
\usage{
heron(a, b, c)
}
\arguments{
\item{a}{Numeric length of side a}

\item{b}{Numeric length of side b}

\item{c}{Numeric length of side c}
}
\value{
The triangle area
}
\description{
Compute triangle area with heron's formula
}
\examples{
heron(3, 4, 5)
heron(6, 8, 10)

}
