% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron.R
\name{heron}
\alias{heron}
\title{Heron}
\usage{
heron(a, b, c)
}
\arguments{
\item{a}{est un cote du trianble (valeur numerique)}

\item{b}{est un cote du trianble (valeur numerique)}

\item{c}{est un cote du trianble (valeur numerique)}
}
\value{
la fonction retourne le resultat de l'aire du triangle (heron)
}
\description{
Heron
}
\examples{
a=4; b=2; c=3
heron(a,b,c)

a=5; b=8; c=7
heron(a,b,c)

}
