% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron.R
\name{heron}
\alias{heron}
\title{heron}
\usage{
heron(a, b, c)
}
\arguments{
\item{a}{Un coté du triangle}

\item{b}{Un second côté du triangle}

\item{c}{Le dernier côté du triangle}
}
\value{
Retourne l'aire du triangle
}
\description{
Quel est le diamètre d'un triangle?
}
\examples{
heron(10,20,15)
heron("bidule","chouette",3)
}
