% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron.R
\name{heron}
\alias{heron}
\title{heron}
\usage{
heron(a, b, c)
}
\arguments{
\item{a}{Longueur du côté a du triangle}

\item{b}{Longueur du côté b du triangle}

\item{c}{Longueur du côté c du triangle}
}
\description{
Fonction heron permettant de calculer la formule du heron, qui permet de calculer
l'air d'un triangle à partir de 3 paramètres : a,b,c
}
\examples{
heron(1,2,3)
heron(4,5,6)
}
