% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron.R
\name{heron}
\alias{heron}
\title{La fonction heron}
\usage{
heron(a, b, c)
}
\arguments{
\item{a}{longueur du premier coté}

\item{b}{longueur du deuxieme coté}

\item{c}{longueur du troisième coté}
}
\value{
Retourne l'aire du triangle.
}
\description{
La fonction permet de calculer l'aire d'un triangle
à partir de  la longueur de ses trois côtés.
}
\examples{
heron(1,2,3) égal à 0
heron(3,4,5) égal à 6
}
