% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron.R
\name{heron}
\alias{heron}
\title{Calculation of the area of a triangle}
\usage{
heron(a, b, c)
}
\arguments{
\item{a}{A number}

\item{b}{A number}

\item{c}{A number}
}
\value{
A number : the area value of a triangle
}
\description{
Calculation of the area of a triangle
}
\examples{
heron(8,10,6)
heron(2,2,1)

}
