% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heron2.R
\name{heron2}
\alias{heron2}
\title{A Leveled-up function calculating the area of a triangle using Heron's formula}
\usage{
heron2(a, b, c)
}
\arguments{
\item{a}{a positive float}

\item{b}{a positive float}

\item{c}{a positive float}
}
\value{
area of a triangle
}
\description{
A Leveled-up function calculating the area of a triangle using Heron's formula
}
\examples{
heron2(3,3,3)
heron2(2,2,2)

}
