% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HERONEpitopeDataSet-class}
\alias{HERONEpitopeDataSet-class}
\alias{.HERONEpitopeDataSet}
\alias{HERONEpitopeDataSet}
\title{HERONEpitopeDataSet object and constructors}
\usage{
HERONEpitopeDataSet(pvalue, ...)
}
\arguments{
\item{pvalue}{calculate epitope p-value matrix}

\item{...}{arguments provided to \code{SummarizedExperiment}, including
metadata}
}
\value{
HERONEpitopeDataSet object
}
\description{
\code{HERONEpitopeDataSet} is a subclass of \code{SummarizedExperiment}
used to hold assay information on the epitope-level
}
\examples{
pval <- matrix(runif(100),ncol=4)
HERONEpitopeDataSet(pvalue = pval)
}
