% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HERONProteinDataSet-class}
\alias{HERONProteinDataSet-class}
\alias{.HERONProteinDataSet}
\alias{HERONProteinDataSet}
\title{HERONProteinDataSet object and constructors}
\usage{
HERONProteinDataSet(pvalue, ...)
}
\description{
TODO
}
