% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HERONSequenceDataSet-class}
\alias{HERONSequenceDataSet-class}
\alias{.HERONSequenceDataSet}
\alias{HERONSequenceDataSet}
\title{HERONSequenceDataSet object and constructors}
\usage{
HERONSequenceDataSet(exprs, ...)
}
\description{
\code{HERONSequenceDataSet} is a subclass of \code{SummarizedExperiment},
used to store the expression values, intermediate calculations, and
results of a differential binding code.
}
