% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{makeCalls}
\alias{makeCalls}
\title{Title}
\usage{
makeCalls(padj_mat, padj_cutoff = 0.05, pData)
}
\arguments{
\item{padj_mat}{adjusted p-value matrix}

\item{padj_cutoff}{adjusted cutoff}

\item{pData}{optional pData data.frame with the condition column defined

 If the condition column is defined, then K of N will also be reported on
 each condition named for the samples.}
}
\value{
list of results
sample => calls made on the sample (column)-level
k_of_n => data.frame with K of N statistics.
}
\description{
Title
}
