% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{makeCalls}
\alias{makeCalls}
\title{Make calls on an input matrix of p-adjusted values}
\usage{
makeCalls(padj_mat, padj_cutoff = 0.05, pData)
}
\arguments{
\item{padj_mat}{adjusted p-value matrix}

\item{padj_cutoff}{adjusted cutoff}

\item{pData}{optional pData data.frame with the condition column defined

 If the condition column is defined, then K of N will also be reported on
 each condition named for the samples.}
}
\value{
list of results
sample => calls made on the sample (column)-level
k_of_n => data.frame with K of N statistics.
}
\description{
This function takes in a matrix of p-values and using
a cutoff, finds the calls on the column/sample sample-level
and calculates the number of samples (K) and the frequency
of samples (F).  If the pData object is provided with a
condition columns defined, then a K and F is calculated for
each condition.
}
\examples{
data(heffron2021_wuhan)
probe_meta <- attr(heffron2021_wuhan, "probe_meta")
pData <- attr(heffron2021_wuhan, "pData")
pval_res <- calcProbePValuesSeqMat(heffron2021_wuhan, probe_meta, pData)
calls_res <- makeCalls(pval_res)
}
