% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{makeEpitopeCalls}
\alias{makeEpitopeCalls}
\title{Make Epitope Calls}
\usage{
makeEpitopeCalls(
  epitope_sample_padj,
  pData,
  padj_cutoff = 0.05,
  one_hit_filter = TRUE
)
}
\arguments{
\item{epitope_sample_padj}{epitope matrix of p-values}

\item{pData}{optional data.frame of design}

\item{padj_cutoff}{p-value cutoff to use}

\item{one_hit_filter}{filter one hit epitopes?}
}
\value{
a list of results
}
\description{
Make Epitope Calls
}
\examples{
data(heffron2021_wuhan)
probe_meta <- attr(heffron2021_wuhan, "probe_meta")
pData <- attr(heffron2021_wuhan, "pData")
pr_pval_res <- calcProbePValuesSeqMat(heffron2021_wuhan, probe_meta, pData)
pr_calls_res <- makeProbeCalls(pr_pval_res)
epi_segments_uniq_res <- findEpitopeSegments(
probe_calls = pr_calls_res,
segment.method = "unique"
)
epi_segments_uniq_res <- findEpitopeSegments(
probe_calls = pr_calls_res,
segment.method = "unique"
)
epi_pval_uniq <- calcEpitopePValuesMat(
probe_pvalues = attr(pr_pval_res, "pvalue"),
epitope_ids = epi_segments_uniq_res,
metap_method = "wilkinsons_max1"
)
epi_padj_uniq <- p_adjust_mat(epi_pval_uniq, method="BH")
makeEpitopeCalls(epi_padj_uniq)
}
