% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{makeEpitopeCallsEDS}
\alias{makeEpitopeCallsEDS}
\title{Title}
\usage{
makeEpitopeCallsEDS(epi_ds, pData, padj_cutoff = 0.05, one_hit_filter = TRUE)
}
\arguments{
\item{one_hit_filter}{}
}
\description{
Title
}
