% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{makeProbeCalls}
\alias{makeProbeCalls}
\title{Making Probe-level Calls}
\usage{
makeProbeCalls(
    probe_sample_padj,
    pData,
    probe_cutoff = 0.05,
    one_hit_filter = TRUE
)
}
\arguments{
\item{probe_sample_padj}{a matrix of adjusted p-values where each row is a
feature and column is a sample}

\item{pData}{optional design matrix for condition calling.}

\item{probe_cutoff}{cutoff to use when calling probes}

\item{one_hit_filter}{Indicator to remove probe hits that do not have a
matching consecutive probe and if the probe is only found in 1 sample}
}
\value{
list of results
sample_probes -> logical matrix of calls on probes
}
\description{
\code{makeProbeCalls} returns call information on a probe matrix that has
been scored by the
}
\examples{
data(heffron2020_wuhan)
probe_meta <- attr(heffron2020_wuhan, "probe_meta")
pData <- attr(heffron2020_wuhan, "pData")
pval_res <- calcProbePValuesSeqMat(heffron2020_wuhan, probe_meta, pData)
calls_res <- makeProbeCalls(pval_res)
}
