% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.R
\name{makeProbeCalls}
\alias{makeProbeCalls}
\title{Making Probe-level Calls}
\usage{
makeProbeCalls(
  probe_sample_padj,
  probe_cutoff = 0.05,
  probes = rownames(probe_sample_padj),
  proteins = getProteinLabel(probes),
  positions = getProteinStart(probes),
  protein_tiling = getProteinTiling(probes),
  one_hit_filter = TRUE
)
}
\arguments{
\item{probe_sample_padj}{a matrix of adjusted p-values where each row is a
feature and column is a sample}

\item{one_hit_filter}{}
}
\value{

}
\description{
\code{makeProbeCalls} returns call information on a probe matrix that has
been scored by the
}
