% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{p_adjust_mat}
\alias{p_adjust_mat}
\title{Title}
\usage{
p_adjust_mat(pvalues_mat, method = "BH")
}
\arguments{
\item{pvalues_mat}{matrix of p-values}

\item{method}{what adjustment algorithm to use (see p.adjust)}
}
\value{
matrix of column-by-column adjusted p-values
}
\description{
Title
}
