% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{quantileNormalize}
\alias{quantileNormalize}
\title{Normalize a matrix/data.frame using quantile normalization}
\usage{
quantileNormalize(in_mat)
}
\arguments{
\item{in_mat}{matrix or data.frame of numeric values to be normalized}
}
\value{
normalized data.frame
}
\description{
Normalize a matrix/data.frame using quantile normalization
}
\examples{
data(heffron2020_wuhan)
seq_mat_qn = quantileNormalize(heffron2020_wuhan)

}
