% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{quantileNormalize}
\alias{quantileNormalize}
\title{Normalize the exprs assay using quantile normalization}
\usage{
quantileNormalize(se)
}
\arguments{
\item{se}{SummarizedExperiment with exprs assay}
}
\value{
SummarizedExperiment with exprs assay normalized
}
\description{
Normalize the exprs assay using quantile normalization
}
\examples{
data(heffron2021_wuhan)
seq_ds_qn <- quantileNormalize(heffron2021_wuhan)
}
